firebase_auth_project/requirements.txt
Flask
gunicorn
firebase-admin
google-cloud-datastore
